style: modify entities folder structure
