+++ src/domain/entities/transaction/transaction.entity.ts
@@ -1,6 +1,7 @@
-import { TransactionType } from '../enums/transaction-type.enum';
+import { TransactionType } from '../../enums/transaction-type.enum';
 import { v4 as uuidv4 } from 'uuid';
-import { CreateTransactionDTO } from 'src/application/dtos/create-transaction.dto';
+import { TransactionStatus } from '../../enums/transaction-status.enum';
+import { CreateTransactionDTO } from '../../../application/dtos/create-transaction.dto';
 
 export abstract class Transaction {
   static transactions: Transaction[] = [];
@@ -10,6 +11,7 @@ export abstract class Transaction {
   createdAt: Date;
   customerId: string;
   accountId: string;
+  status: TransactionStatus;
 
   constructor(createTransactionDto: CreateTransactionDTO) {
     this.id = uuidv4();
@@ -18,5 +20,6 @@ export abstract class Transaction {
     this.amount = createTransactionDto.amount;
     this.accountId = createTransactionDto.accountId;
     this.createdAt = new Date();
+    this.status = TransactionStatus.PROCESSED;
   }
 }
